Delay update of PVT plot until it is visible
