+++ ApplicationLibCode/UserInterface/CMakeLists_files.cmake
@@ -23,6 +23,7 @@ set(SOURCE_GROUP_HEADER_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RiuPropertyViewTabWidget.h
     ${CMAKE_CURRENT_LIST_DIR}/RiuPvtPlotPanel.h
     ${CMAKE_CURRENT_LIST_DIR}/RiuPvtPlotUpdater.h
+    ${CMAKE_CURRENT_LIST_DIR}/RiuPvtPlotWidget.h
     ${CMAKE_CURRENT_LIST_DIR}/RiuQwtLinearScaleEngine.h
     ${CMAKE_CURRENT_LIST_DIR}/RiuQwtScalePicker.h
     ${CMAKE_CURRENT_LIST_DIR}/RiuQwtCurvePointTracker.h
@@ -137,6 +138,7 @@ set(SOURCE_GROUP_SOURCE_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RiuPropertyViewTabWidget.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RiuPvtPlotPanel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RiuPvtPlotUpdater.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/RiuPvtPlotWidget.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RiuQwtLinearScaleEngine.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RiuQwtScalePicker.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RiuQwtCurvePointTracker.cpp
